requirements.txt
boto3==1.34.144
botocore==1.34.144
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
idna==3.7
jmespath==1.0.1
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.1
mdurl==0.1.2
Pygments==2.18.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
shellingham==1.5.4
six==1.16.0
textual==0.71.0
typer==0.12.3
typing_extensions==4.12.2
uc-micro-py==1.0.3
urllib3==2.2.1
